requirements.txt
Pillow
tqdm
aiohttp
requests
